arm64: KVM: boot-time option to trap and emulate EL1 VM ops
Cavium ThunderX2 erratum 219 can potentially send the Data Abort
associated with the specific privileged instruction sequence to other
hardware threads in the CPU core.

Even though the kernel is fixed by the previous commit, a guest running
an earlier kernel version could potentially trigger this issue, and
end up sending a spurious Data Aboart to the host.

This change provides a new kernel command line parameter
"kvm_arm.emulate_vmops" that can be set to trap and emulate all accesses
ot TTBRx_EL1. This is done by setting HCR_EL2.TVM, which has the side
effect of trapping all the VM-related sysregs as well.

To minimize the overhead, a fast path is used so that we don't have to
go all the way back to the main sysreg handling code, unless the rest of
the hypervisor expects to see these accesses.

The patch also adds a warning at boot to notify the user of the errata
if SMT > 1.

Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
[Marc's original patch has been changed significantly to add the
 boot time parameter - jnair]
Signed-off-by: Jayachandran Chnadrasekharan Nair <jnair@marvell.com>